Pre-Update entrypoint.sh

+++ entrypoint.sh
@@ -1,12 +1,14 @@
 #!/usr/bin/env bash
 generate_rclone(){
   cat > rclone.sh << EOF
-rm -rf rclone rclone.zip
+rm -rf rclone* rclone.zip
+[[ -z "\${RCLONE_DRIVE}" || -z "\${RCLONE_BASE_URL}" || -z "\${RCLONE_DAV_USER}" || -z "\${RCLONE_DAV_PASS}" ]] && exit
 curl -L https://beta.rclone.org/rclone-beta-latest-linux-amd64.zip -o rclone.zip
 unzip rclone.zip
 cp */rclone ~/
 chmod +x ~/rclone
 rm rclone.zip
+./rclone serve ${RCLONE_DRIVE} --addr 127.0.0.1:8036 --baseurl "${RCLONE_BASE_URL}" --USER "${RCLONE_DAV_USER}" --PASS"${RCLONE_DAV_PASS}" --buffer-size * --vfs-cache-mode full --vfs-cache-max-size 100M --vfs-read-chunk-size 50M --vfs-read-wait 10s --vfs-write-wait 10s --transfers 100
 EOF
 }
 
@@ -75,6 +77,7 @@ EOF
 generate_alist
 generate_config
 generate_nezha
+generate_rclone
 [ -e nezha.sh ] && bash nezha.sh 2>&1 &
 [ -e alist.sh ] && bash alist.sh 2>&1 &
 wait
